updated code

+++ controllers/API/v1/post_api.js
@@ -1,35 +1,44 @@
-const Post=require("../../../models/post")
-const Comment=require("../../../models/comment")
-// find  all post 
-module.exports.index=async(req,res)=>{
-          // Populate the user and comments for each post
-          const posts = await Post.find({})
-            .sort("-createdAt")
-            .populate("user")
-            .populate({
-              path: "comments",
-              populate: {
-                path: "user",
-              },
-            })
-    return res.status(200).json({
-        message:"lists of posts",
-        posts:posts
-    })
-}
+const Post = require("../../../models/post");
+const Comment = require("../../../models/comment");
+
+module.exports.index = async function (req, res) {
+  const posts = await Post.find({})
+    .sort("-createdAt")
+    .populate("user", "-password") //ignoring the password
+    .populate({
+      path: "comments",
+      populate: {
+        path: "user",
+      },
+    });
+
+  return res.json(200, {
+    message: "List of posts",
+    post: posts,
+  });
+};
+
 module.exports.destroy = async (req, res) => {
   try {
-    const postId = req.params.id.trim(); // Sanitize the id
-    const post = await Post.findById(postId);
-    console.log(post);
-    await post.deleteOne();
-    
-    // Delete comments associated with the post
-    await Comment.deleteMany({ post: postId });
-    
-    return res.status(200).json({ message: 'post and associates are deleted' });
+    const post = await Post.findById(req.params.id);
+    if (!post) {
+      return res.status(404).send({ message: "Post not found" });
+    }
+    if (post.user.toString() === req.user.id) {
+      post.deleteOne();
+
+      // Delete comments associated with the post
+      await Comment.deleteMany({ post: req.params.id });
+      return res.json(200, {
+        message: "Post and associated comments are deleted successfully!",
+      });
+    } else {
+      return res
+        .status(401)
+        .send({ message: "You're not authorized to delete this post" });
+    }
   } catch (err) {
-    console.log(err);
-    return res.status(500).send("Internal server error");
+    console.error(err);
+    return res.status(500).send({ error: err.message });
   }
 };
